add dynamic search for absorber and donor view
Aborber is the hit-test view of the drop point, closest in the view hierarchy,
that is also designated as an absorber. The donor is the closest ancestor
in the view hierarchy that is designated a donor.

Still need to decide how to handle designating a view as a donor or
absorber.