refactor. use assocObjects instead of dictionaries to track delegates.
The DnD server used to hold two dictionaries from UIViews to their delegates.
This had two problems. First, it created a strong reference to the delegates, when
the "views" should be holding weak references. Second, there was no obvious
way to clear dead entries in the dictionaries when the views themselves were
deallocated. Associated objects solves both problems and is terser.